+++ packages/realm/src/Realm.ts
@@ -48,10 +48,12 @@ import {
   SubscriptionSet,
   SyncSession,
   TypeAssertionError,
+  TypeHelpers,
   Unmanaged,
   UpdateMode,
   assert,
   binding,
+  createResultsAccessor,
   defaultLogger,
   defaultLoggerLevel,
   extendDebug,
@@ -938,25 +940,27 @@ export class Realm {
   objects<T = DefaultObject>(type: string): Results<RealmObject<T> & T>;
   objects<T extends AnyRealmObject = RealmObject & DefaultObject>(type: Constructor<T>): Results<T>;
   objects<T extends AnyRealmObject>(type: string | Constructor<T>): Results<T> {
-    const { objectSchema, wrapObject } = this.classes.getHelpers(type);
+    const { internal, classes } = this;
+    const { objectSchema, wrapObject } = classes.getHelpers(type);
     if (isEmbedded(objectSchema)) {
       throw new Error("You cannot query an embedded object.");
     } else if (isAsymmetric(objectSchema)) {
       throw new Error("You cannot query an asymmetric object.");
     }
 
-    const table = binding.Helpers.getTable(this.internal, objectSchema.tableKey);
-    const results = binding.Results.fromTable(this.internal, table);
-    return new Results<T>(this, results, {
-      get(results: binding.Results, index: number) {
-        return results.getObj(index);
+    const table = binding.Helpers.getTable(internal, objectSchema.tableKey);
+    const results = binding.Results.fromTable(internal, table);
+    const typeHelpers: TypeHelpers<T> = {
+      fromBinding(value) {
+        return wrapObject(value as binding.Obj) as T;
       },
-      fromBinding: wrapObject,
-      toBinding(value: unknown) {
+      toBinding(value) {
         assert.instanceOf(value, RealmObject);
         return value[INTERNAL];
       },
-    });
+    };
+    const accessor = createResultsAccessor<T>({ realm: this, typeHelpers, itemType: binding.PropertyType.Object });
+    return new Results<T>(this, results, accessor, typeHelpers);
   }
 
   /**